Broke out normalizeNestedObjects logic to separate file
